+++ packages/ilmomasiina-client/src/modules/editSignup/index.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren, useMemo } from "react";
 
 import type { SignupForEditResponse, SignupUpdateResponse } from "@tietokilta/ilmomasiina-models";
-import { EDIT_TOKEN_HEADER } from "@tietokilta/ilmomasiina-models";
+import { EDIT_TOKEN_HEADER, SignupPaymentStatus } from "@tietokilta/ilmomasiina-models";
 import { ApiError, apiFetch } from "../../api";
 import { useAbortablePromise } from "../../utils/abortable";
 import { getLocalizedEvent, getLocalizedSignup } from "../../utils/localizedEvent";
@@ -53,14 +53,30 @@ export function useEditSignupState({ id, editToken, paid, language }: EditSignup
           },
         });
       }
+
+      const isNew = !response.signup.confirmed;
+      const alreadyPaid = response.signup.paymentStatus === SignupPaymentStatus.PAID;
+
       const now = Date.now();
+      const editingClosedOnLoad = response.signup.editableForMillis === 0;
+
       return {
         ...response,
         paymentError,
         // Compute these once when the response arrives.
-        editingClosedOnLoad: response.signup.editableForMillis === 0,
+        editingClosedOnLoad,
         confirmableUntil: now + response.signup.confirmableForMillis,
         editableUntil: now + response.signup.editableForMillis,
+        isNew,
+
+        // Show payment if there's a price to pay.
+        showPayment: response.signup.price != null && response.signup.price > 0,
+        // Allow editing for non-admins if not closed.
+        canEdit: !editingClosedOnLoad,
+        // Allow name and email editing for non-admins if canEdit and the signup is not confirmed.
+        canEditNameAndEmail: !editingClosedOnLoad && isNew,
+        // Allow editing of paid questions only if canEdit and not already paid.
+        canEditPaidQuestions: !editingClosedOnLoad && !alreadyPaid,
 
         updateSignup: (update: SignupUpdateResponse) => {
           // Only store the update if the hook hasn't unmounted.
@@ -93,7 +109,6 @@ export function useEditSignupState({ id, editToken, paid, language }: EditSignup
     ...result,
     localizedEvent,
     localizedSignup,
-    isNew: result && !result.signup.confirmed,
   });
 }