+++ packages/ilmomasiina-frontend/src/routes/EditSignup/components/QuestionFields.tsx
@@ -18,10 +18,10 @@ import useFieldErrors from "./fieldError";
 type QuestionFieldProps = {
   name: string;
   question: Question;
-  disabled?: boolean;
+  validate?: boolean;
 };
 
-const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
+const QuestionField = ({ name, question, validate = true }: QuestionFieldProps) => {
   const {
     input: { value, onChange },
     meta: { invalid },
@@ -29,16 +29,23 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
   const currentAnswerString = stringifyAnswer(value);
   const currentAnswerArray = useMemo(() => (Array.isArray(value) ? value : []), [value]);
 
+  const { canEdit, canEditPaidQuestions, signup } = useEditSignupContext();
   const { t } = useTranslation();
   const formatError = useFieldErrors();
+
+  // For admins, make all questions optional.
+  // (All of them are editable by the user, and the backend doesn't care for admins.)
+  const isRequired = validate && question.required;
 
   const formatPrice = usePriceFormatter();
   // Show the prices for each option if the question has some paid options.
   // Add a + sign if the signup has a "base price" from the quota.
-  const quotaHasPrice = useEditSignupContext().signup!.quota.price > 0;
+  const quotaHasPrice = signup!.quota.price > 0;
   const hasPrices = questionHasPrices(question);
   const formatOptionPrice = (price?: number) =>
     hasPrices && price != null ? ` (${quotaHasPrice ? "+" : ""}${formatPrice(price)})` : "";
+
+  const disabled = !canEdit || (!canEditPaidQuestions && questionHasPrices(question));
 
   // We need to wrap onChange, as react-final-form complains if we pass radios to it without type="radio".
   // If we pass type="radio", it doesn't provide us with the value of the field.
@@ -52,7 +59,13 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
     onChange(newAnswers);
   });
 
-  const help = question.public ? t("editSignup.publicQuestion") : null;
+  const help =
+    // eslint-disable-next-line no-nested-ternary
+    canEdit && disabled // implies question is uneditable because of !canEditPaidQuestions
+      ? t("editSignup.uneditablePaidQuestion")
+      : question.public
+        ? t("editSignup.publicQuestion")
+        : null;
 
   let input: ReactNode;
   let isCheckboxes = false;
@@ -62,7 +75,7 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
         <Form.Control
           type="text"
           maxLength={250}
-          required={question.required}
+          required={isRequired}
           readOnly={disabled}
           value={currentAnswerString}
           onChange={onFieldChange}
@@ -74,7 +87,7 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
       input = (
         <Form.Control
           type="number"
-          required={question.required}
+          required={isRequired}
           readOnly={disabled}
           value={currentAnswerString}
           onChange={onFieldChange}
@@ -91,7 +104,7 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
           id={`question-${question.id}-option-${optIndex}`}
           value={option}
           label={`${option}${formatOptionPrice(question.prices?.[optIndex])}`}
-          required={question.required && !currentAnswerArray.some((answer) => answer !== option)}
+          required={isRequired && !currentAnswerArray.some((answer) => answer !== option)}
           disabled={disabled}
           checked={currentAnswerArray.includes(option)}
           onChange={onCheckboxChange}
@@ -108,7 +121,7 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
           rows={3}
           cols={40}
           maxLength={250}
-          required={question.required}
+          required={isRequired}
           readOnly={disabled}
           value={currentAnswerString}
           onChange={onFieldChange}
@@ -120,13 +133,13 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
       if (question.options && question.options.length > 3) {
         input = (
           <Form.Select
-            required={question.required}
+            required={isRequired}
             disabled={disabled}
             value={currentAnswerString}
             onChange={onFieldChange}
             isInvalid={invalid}
           >
-            <option value="" disabled={question.required}>
+            <option value="" disabled={isRequired}>
               {t("editSignup.fields.select.placeholder")}
             </option>
             {question.options?.map((option, optIndex) => (
@@ -148,7 +161,7 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
             inline
             value={option}
             label={`${option}${formatOptionPrice(question.prices?.[optIndex])}`}
-            required={question.required}
+            required={isRequired}
             disabled={disabled}
             checked={currentAnswerString === option}
             onChange={onFieldChange}
@@ -167,6 +180,7 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
       key={question.id}
       name={`${name}.${question.id}`}
       label={question.question}
+      // Show required indicator even for admins for information purposes.
       required={question.required}
       help={help}
       checkAlign={isCheckboxes}
@@ -179,15 +193,15 @@ const QuestionField = ({ name, question, disabled }: QuestionFieldProps) => {
 
 type Props = {
   name: string;
-  canEdit: boolean;
+  validate?: boolean;
 };
 
-const QuestionFields = ({ name, canEdit }: Props) => {
+const QuestionFields = ({ name, validate = true }: Props) => {
   const { localizedEvent: event } = useEditSignupContext();
   return (
     <>
       {event!.questions.map((question) => (
-        <QuestionField key={question.id} name={name} question={question} disabled={!canEdit} />
+        <QuestionField key={question.id} name={name} question={question} validate={validate} />
       ))}
     </>
   );