+++ packages/ilmomasiina-frontend/src/routes/Editor/components/EditSignupModal.tsx
@@ -57,8 +57,8 @@ const EditSignupModalBody = ({ handleSubmit, submitting }: FormRenderProps<Signu
       </Modal.Header>
       <Modal.Body>
         {isNew && <FieldRow name="quotaId" as={QuotaField} label={t("editor.editSignup.quota")} required />}
-        <CommonFields canEdit canEditNameAndEmail />
-        <QuestionFields name="answers" canEdit />
+        <CommonFields />
+        <QuestionFields name="answers" validate={false} />
         <FieldRow
           name="sendEmail"
           as={BsForm.Check}
@@ -132,6 +132,10 @@ const EditSignupModal = () => {
       editingClosedOnLoad: false,
       confirmableUntil: Infinity,
       editableUntil: Infinity,
+      showPayment: false,
+      canEdit: true,
+      canEditNameAndEmail: true,
+      canEditPaidQuestions: true,
     };
   }, [values, editedSignup]);